# implementation_examples/openapi.yaml
openapi: 3.1.0
...
info:
  title: Oolix API
  version: 0.2.0
  description: |
    Privacy-safe partner media activation control plane.

    **The rule every endpoint serves.** Oolix is the control plane; Data
    Partners remain the customer-data owners (§1). No request or response in
    this contract carries a `partner_user_id`, a segment membership list or an
    exact audience count. Reach is always a bucket (§72) and attribution is
    always an opaque token resolved server-side by SHA-256 (§90).
  contact:
    name: Oolix engineering
  license:
    name: Oolix proprietary
    identifier: LicenseRef-Oolix-Proprietary
servers:
  - url: http://localhost:4000
    description: Local development (docker compose stack)
  - url: '{scheme}://{host}'
    description: |
      Deployed environments. The host is supplied per environment rather than
      hard-coded, so a generated client can never default to someone else's
      domain.
    variables:
      scheme: { default: https, enum: [https] }
      host: { default: localhost:4000 }
...
security:
  - oidcBearer: []
paths:
...
  /v1/campaigns:
    get:
      operationId: listCampaigns
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Campaign]
      summary: List campaigns
      responses:
        '200': { description: Campaigns }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
    post:
      operationId: createCampaign
      tags: [Campaign]
      summary: Create a campaign draft
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCampaign' }
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Campaign' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/campaigns/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      operationId: getCampaign
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Campaign]
      summary: Campaign detail with per-Partner activations
      description: |
        §42: the parent campaign is an aggregate **view**. Each Partner request
        and each activation carries its own independent state and reason;
        `derived_status` is a projection over the activations.
      responses:
        '200': { description: Campaign detail }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
    patch:
      operationId: updateCampaign
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Campaign]
      summary: Update a draft campaign
      responses:
        '200': { description: Updated }
        '409': { description: 'Campaign is not DRAFT (§76)' }

        '429': { $ref: '#/components/responses/RateLimited' }
...
  /v1/campaigns/{id}/submit:
    post:
      operationId: submitCampaign
      tags: [Campaign, Approval]
      summary: Submit the campaign for Partner review
      description: |
        §40.9 freezes the request version; §101 starts each Partner's own
        review clock. **Idempotency-Key is required** (§53): a retried submit
        must not start a second review.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      responses:
        '200': { description: Submitted }
        '400': { description: 'Idempotency-Key missing (§53)' }

        '429': { $ref: '#/components/responses/RateLimited' }
  # ---------------------------------------------------------------------------
...
  /v1/partner-requests:
    get:
      operationId: listPartnerRequests
      tags: [Approval]
      summary: The Partner's review queue
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: status, in: query, schema: { $ref: '#/components/schemas/PartnerRequestStatus' } }
      responses:
        '200': { description: Requests awaiting decision }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/partner-requests/{id}:
    get:
      operationId: getPartnerRequest
      tags: [Approval]
      summary: Everything the Partner must see before deciding
      description: |
        §41's checklist in one response: Buyer identity, category, segment,
        channel and placement, frequency cap, creative preview bound to its
        content hash, landing destination, commercial basis, lead definition,
        expansion flag and the policy version the decision will bind to.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Review detail }
        '404': { $ref: '#/components/responses/NotFound' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/partner-requests/{id}/approve:
    post:
      operationId: approvePartnerRequest
      tags: [Approval]
      summary: Approve a Partner request
      description: |
        §31: the Data Partner makes the **final** decision; neither the network
        sponsor nor Oolix can override it. §66.2: the user who created the
        request may not approve it, even within one organization.

        On success this creates one activation per approved channel (§13) and
        signs a manifest for each owned-media activation (§75). An external
        channel activation stays `PENDING_CHANNEL_CHECK` with no manifest --
        approval alone must never enable an upload (§15, §16).

        **Idempotency-Key is required** (§53).
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Approve' }
      responses:
        '201': { description: Approved; activations created }
        '400': { description: 'Idempotency-Key missing (§53)' }
        '403': { description: 'Wrong role, or self-approval (§66, §66.2)' }
        '409': { description: 'Invalid state transition, or request expired (§76, §101)' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/partner-requests/{id}/request-change:
    post:
      operationId: requestPartnerRequestChange
      tags: [Approval]
      summary: Ask the Buyer to change the request
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Change requested }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/partner-requests/{id}/reject:
    post:
      operationId: rejectPartnerRequest
      tags: [Approval]
      summary: Reject a Partner request
      description: '§76: a REJECTED request never becomes APPROVED without a new version.'
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Rejected }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/partner-requests/{id}/revoke:
    post:
      operationId: revokePartnerRequest
      tags: [Approval]
      summary: Revoke an approved or live request
      description: |
        §57: local serving stops immediately and external cleanup is queued.
        The manifest is revoked so an Agent holding a cached copy drops it on
        its next control sync (§75).
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Revoked; activations ENDING }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/partner-requests/{id}/extend:
    post:
      operationId: extendPartnerRequestReview
      tags: [Approval]
      summary: Extend the review window once
      description: '§101: one extension of up to 7 days, with a mandatory audited reason.'
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Extended }
        '409': { description: 'Already extended once (§101)' }

        '429': { $ref: '#/components/responses/RateLimited' }
  # ---------------------------------------------------------------------------
...
  /v1/partner/agents:
    get:
      operationId: listAgents
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Agent]
      summary: List this Partner's Agents
      responses:
        '200': { description: Agents with heartbeat and config age }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/partner/agents/bootstrap-tokens:
    post:
      operationId: createAgentBootstrapToken
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Agent]
      summary: Mint a single-use Agent bootstrap token
      description: |
        §92.1: 32 random bytes, single-use, 15 minutes. Returned in plaintext
        **once**; only its SHA-256 is stored. PARTNER_SECURITY_ADMIN only --
        §66 separates this from the commercial PARTNER_ADMIN.
      responses:
        '201':
          description: Bootstrap token (shown once)
        '403': { $ref: '#/components/responses/Forbidden' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/partner/agents/bootstrap-tokens/revoke:
    post:
      operationId: revokeAgentBootstrapTokens
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Agent]
      summary: Revoke all unused bootstrap tokens
      responses:
        '201': { description: Revoked }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/partner/agents/{id}/revoke:
    post:
      operationId: revokeAgent
      tags: [Agent]
      summary: Revoke an Agent immediately
      description: |
        §69.3: "Oolix may revoke an Agent immediately during incident
        response." Effective at once -- the guard re-checks Agent status on
        every call rather than waiting for the 15-minute token to expire.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Revoked }
        '400': { description: 'A reason is required for the audit trail (§83)' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /agent/v1/register:
    post:
      operationId: agentRegister
      tags: [Agent]
      summary: Register a Partner Agent
      description: |
        §92.2. Public because the Agent has no access token yet; it
        authenticates with the single-use bootstrap token in the body.

        The Agent supplies only its **public** JWK -- §92.2 keeps the private
        key inside Partner infrastructure, which is what makes a signed report
        batch meaningful tamper evidence.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterAgent' }
      responses:
        '201':
          description: Agent identity and endpoints
        '401': { description: 'Bootstrap token invalid, used or expired (§92.1)' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /agent/v1/token:
    post:
      operationId: agentIssueToken
      tags: [Agent]
      summary: Exchange a client assertion for an access token
      description: |
        §92.3. The ES256 client assertion **is** the credential, verified
        against the public key recorded at registration. Tokens live 15
        minutes.
      security: []
      responses:
        '200':
          description: Access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  token_type: { type: string, const: Bearer }
                  expires_in: { type: integer, const: 900 }
        '401': { description: Assertion verification failed }

        '429': { $ref: '#/components/responses/RateLimited' }
  /agent/v1/heartbeat:
    post:
      operationId: agentHeartbeat
      tags: [Agent]
      summary: Agent heartbeat
      description: '§52.4. Feeds the §78.2 staleness alerts.'
      security:
        - agentBearer: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '201': { description: 'Acknowledged, with config freshness' }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        §64: OIDC access token. Identity is delegated to the provider; Oolix
        stores only the subject claim. Roles are read from the database for the
        active organization, never from a token claim (§4.2).
    agentBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '§92.3: Oolix-issued Agent access token, 15-minute lifetime.'
    crmApiKey:
      type: http
      scheme: bearer
      description: "§71: the Buyer's CRM API key. Hashed at rest and rotatable."
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 200 }
      description: |
        §22.3, §99. Same key + same payload replays the stored result; same key
        + different payload is `409 IDEMPOTENCY_CONFLICT`. **Required** on
        campaign submit, approval and mark-paid (§53).
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, maxLength: 200 }
      description: |
        §53 makes the key MANDATORY here. These three operations each have an
        effect that must not happen twice -- starting a second Partner review,
        creating a second set of activations, or paying a payout twice -- so a
        retry without a key is refused rather than guessed at.
    AgentId:
      name: X-Agent-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: |
        §92.4: must match the access token's subject. Without this check a
        valid token for one Agent could be replayed while claiming to be
        another.
    OrgId:
      name: X-Org-Id
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: |
        Selects which organization the request acts within, for a user who
        belongs to several (§34 role switching without a second account).
...
  responses:
    Unauthorized:
      description: '`AUTH_001` -- invalid or expired token.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: '`PERM_001` / `PERM_002` -- permission or organization scope denied.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: |
        `CAMP_001` / `PART_001`. Also returned where an object exists but is
        not visible to the caller, so existence cannot be probed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: '`VAL_001` -- schema validation failed.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: |
        `SYS_001` -- §86's per-class budget for this caller is exhausted. The
        response carries `X-RateLimit-*`, `Retry-After` and
        `retry_after_seconds`, so a client can pace itself instead of guessing
        (§94).
      headers:
        X-RateLimit-Limit:
          { schema: { type: integer }, description: Requests allowed in the window. }
        X-RateLimit-Remaining:
          { schema: { type: integer }, description: Requests left in the window. }
        X-RateLimit-Reset:
          { schema: { type: integer }, description: Unix seconds when the window frees up. }
        Retry-After: { schema: { type: integer }, description: Seconds to wait before retrying. }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    IdempotencyConflict:
      description: '`IDEMPOTENCY_CONFLICT` -- key reused with a different payload (§99).'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [error]
      description: |
        §53, §99. Every error uses this envelope: a canonical code, a human
        message, an explicit retryable flag, field errors and the correlation
        id. §99 forbids exposing stack traces or secrets.
      properties:
        error:
          type: object
          required: [code, message, retryable, correlation_id, field_errors]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message: { type: string }
            retryable: { type: boolean }
            correlation_id: { type: string }
            retry_after_seconds: { type: integer }
            field_errors:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
    ErrorCode:
      type: string
      description: '§77.2 canonical error registry.'
      enum:
        [
          AUTH_001,
          PERM_001,
          PERM_002,
          VAL_001,
          CAMP_001,
          CAMP_002,
          CAMP_003,
          PART_001,
          PART_002,
          CHAN_001,
          CHAN_002,
          AGENT_001,
          SYS_001,
          SYS_002,
          IDEMPOTENCY_CONFLICT,
        ]

    # --- §89 canonical channel enum ------------------------------------------
    Channel:
      type: string
      description: |
        §89: exactly these four strings in the database, APIs, manifests,
        events and frontend state. Eligibility is a separate state, never
        encoded into the channel constant.
      enum: [PARTNER_WEB, PARTNER_APP, META, GOOGLE]
...
    Money:
      type: object
      description: '§53, §73: integer minor units plus ISO-4217 code. Never a float.'
      required: [amount_minor, currency]
      properties:
        amount_minor: { type: integer, format: int64, minimum: 1 }
        currency: { type: string, minLength: 3, maxLength: 3 }
...
    PartnerRequestStatus:
      type: string
      description: '§76 + §101. `PARTNER_REVIEW` is canonical; `PENDING_APPROVAL` does not exist.'
      enum: [DRAFT, PARTNER_REVIEW, CHANGE_REQUESTED, APPROVED, REJECTED, REVOKED, EXPIRED]
    ActivationStatus:
      type: string
      enum: [PENDING_CHANNEL_CHECK, READY, SYNCING, LIVE, PAUSED, FAILED, ENDING, ENDED]

    # --- request bodies ------------------------------------------------------
...
    CreateCampaign:
      type: object
      required:
        [name, objective, brand_id, category, purpose_id, budget, start_at, end_at, geographies]
      properties:
        name: { type: string, minLength: 3, maxLength: 120 }
        objective: { type: string, enum: [QUALIFIED_LEADS, CONVERSIONS, CLICKS, AWARENESS] }
        brand_id: { type: string, format: uuid }
        category: { type: string }
        purpose_id:
          type: string
          description: '§81: explicit and versioned -- never a permanent "eligible" boolean.'
        budget: { $ref: '#/components/schemas/Money' }
        start_at: { type: string, format: date-time }
        end_at: { type: string, format: date-time }
        geographies: { type: array, items: { type: string }, minItems: 1 }
        landing_url:
          type: string
          format: uri
          description: |
            §40.2, §40.7: HTTPS, on the brand's verified domain, and carrying no
            PII query parameters. Required for click, lead and conversion
            objectives.
        lead_definition:
          $ref: '#/components/schemas/LeadDefinition'
    LeadDefinition:
      type: object
      description: |
        §40.8. Required for outcome objectives: §50 forbids paying a Partner on
        unverified clicks, so a CPQL campaign that never defines "qualified"
        has no settleable basis.
      required: [qualified_statuses]
      properties:
        qualified_statuses:
          type: array
          items: { type: string, enum: [VALID, QUALIFIED, CONVERTED] }
          minItems: 1
        duplicate_window_days: { type: integer, minimum: 1, maximum: 365, default: 30 }
        valid_lead_rule: { type: string }
        qualified_lead_rule: { type: string }
        conversion_event: { type: string }
    Campaign:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        objective: { type: string }
        status: { type: string }
        derived_status:
          type: string
          description: '§42: a projection over the activations, never a stored source of truth.'
        budget: { $ref: '#/components/schemas/Money' }
        version: { type: integer }
        activations:
          type: array
          description: |
            §42: each activation carries its OWN state. A campaign is routinely
            LIVE at one Partner and REJECTED at another, and flattening that
            into a single status would hide the rejection.
          items:
            type: object
            properties:
              activation_id: { type: string, format: uuid }
              partner_org_id: { type: string, format: uuid }
              channel: { $ref: '#/components/schemas/Channel' }
              status: { $ref: '#/components/schemas/ActivationStatus' }
              status_reason: { type: [string, 'null'] }
...
    Approve:
      type: object
      required: [approved_channels, approved_creative_version_ids]
      description: '§41: the Partner may approve a SUBSET of what was requested.'
      properties:
        decision_version: { type: integer, minimum: 1, default: 1 }
        approved_channels:
          { type: array, items: { $ref: '#/components/schemas/Channel' }, minItems: 1 }
        approved_placement_ids: { type: array, items: { type: string, format: uuid } }
        approved_budget_minor: { type: integer, minimum: 1 }
        approved_creative_version_ids:
          type: array
          items: { type: string, format: uuid }
          minItems: 1
          description: '§70: the Partner approves specific creative VERSIONS, bound to a content hash.'
        audience_expansion_allowed: { type: boolean, default: false }
        partner_payout:
          type: object
          properties:
            model: { type: string, enum: [CPM, CPC, CPL, CPQL, FIXED, HYBRID] }
            unit_price_minor: { type: integer, minimum: 0 }
        approval_note: { type: string, maxLength: 1000 }
...
    RegisterAgent:
      type: object
      required: [bootstrap_token, agent_public_key_jwk, agent_version, capabilities]
      properties:
        bootstrap_token: { type: string, minLength: 20 }
        agent_public_key_jwk:
          type: object
          required: [kty, crv, x, y]
          description: |
            **Public** key only. A `d` parameter is rejected: its presence means
            the Agent is leaking its own private key (§92.2).
          properties:
            kty: { type: string, const: EC }
            crv: { type: string, const: P-256 }
            x: { type: string }
            y: { type: string }
        agent_version: { type: string, maxLength: 32 }
        capabilities: { type: array, items: { $ref: '#/components/schemas/Channel' }, minItems: 1 }
...